% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.support_operations.R
\name{describe_severity_levels}
\alias{describe_severity_levels}
\title{Returns the list of severity levels that you can assign to an AWS Support case}
\usage{
describe_severity_levels(
  language = "string"
)
}
\arguments{
\item{language}{The ISO 639-1 code for the language in which AWS provides support. AWS Support currently supports English ("en") and Japanese ("ja"). Language parameters must be passed explicitly for operations that take them.}
}
\description{
Returns the list of severity levels that you can assign to an AWS Support case. The severity level for a case is also a field in the CaseDetails data type included in any CreateCase request.
}
\examples{

}
