% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.support_operations.R
\name{describe_trusted_advisor_check_refresh_statuses}
\alias{describe_trusted_advisor_check_refresh_statuses}
\title{Returns the refresh status of the Trusted Advisor checks that have the specified check IDs}
\usage{
describe_trusted_advisor_check_refresh_statuses(checkIds)
}
\arguments{
\item{checkIds}{The IDs of the Trusted Advisor checks to get the status of. \strong{Note:} Specifying the check ID of a check that is automatically refreshed causes an \code{InvalidParameterValue} error.}
}
\description{
Returns the refresh status of the Trusted Advisor checks that have the specified check IDs. Check IDs can be obtained by calling DescribeTrustedAdvisorChecks.
}
\details{
Some checks are refreshed automatically, and their refresh statuses cannot be retrieved by using this operation. Use of the \code{DescribeTrustedAdvisorCheckRefreshStatuses} operation for these checks causes an \code{InvalidParameterValue} error.
}
\examples{

}
