% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.support_operations.R
\name{describe_trusted_advisor_check_summaries}
\alias{describe_trusted_advisor_check_summaries}
\title{Returns the summaries of the results of the Trusted Advisor checks that have the specified check IDs}
\usage{
describe_trusted_advisor_check_summaries(checkIds)
}
\arguments{
\item{checkIds}{The IDs of the Trusted Advisor checks.}
}
\description{
Returns the summaries of the results of the Trusted Advisor checks that have the specified check IDs. Check IDs can be obtained by calling DescribeTrustedAdvisorChecks.
}
\details{
The response contains an array of TrustedAdvisorCheckSummary objects.
}
\examples{

}
